Revert to xarray.DataArray.plot for plotting in NCL_conOncon_1.py
This reverts commit e1278572ef7adcb3125658e693790a5a5432a089

Previous attempts at getting rid of a warning resulted in a broken plot,
this commit fixes the broken plot and suppresses the warning message.

When using xarray.DataArray.plot(), "seaborn.apionly" is imported and
throws a UserWarning. This only seems to happen on Read the Docs, not
when building docs locally. This warning is now ignored globally for all
examples in conf.py.