Fix xarray.DataArray keyword args in mask_1.py
The xr_add_cyclic function defined in mask_1.py includes the 'encoding'
keyword argument to the xarray.DataArray constructor that was deprecated
as of Xarray 0.14 and was removed in 0.15.

I believe the xr_add_cyclic function will eventually go into geocat-viz,
so we should make this change there as well.